requirements.txt
django
python-dotenv
langchain
langchain-google-genai
PyGithub
